requirements.txt
fastapi
uvicorn
torch
torchvision
ftfy
regex
tqdm
Pillow
requests
git+https://github.com/openai/CLIP.git
